Integer, Big Integerla güncellendi. Company ve Department entityleri oluşturuldu. Role işlemi enumeration üzerinden kullanmak yerine entity oluşturulup yeniden üretilebilecek şekilde tekrardan tasarlandı.

+++ SCM.Application/Models/DTOs/Approves/ApproveDTO.cs
@@ -1,11 +1,12 @@
 ﻿using SCM.Domain.Entities;
+using System.Numerics;
 
 namespace SCM.Application.Models.DTOs.Approves
 {
     public class ApproveDTO
     {
-        public int Id { get; set; }
-        public int RequestId { get; set; }
+        public BigInteger Id { get; set; }
+        public BigInteger RequestId { get; set; }
         public bool IsApproved { get; set; }
         public decimal ApprovedAmount { get; set; }
         public RequestStatus Status { get; set; }
